requirements.txt
fastapi>=0.115.0
pgvector>=0.4.1
psycopg[binary,pool]>=3.2.3
langgraph>=0.0.10
langchain-ollama>=0.0.1
mcp>=1.1.0
numpy<2.0
openai>=1.55.0
psutil>=6.1.0
pypdf>=5.1.0
python-docx>=1.1.2
python-dotenv>=1.0.1
requests>=2.32.3
sentence-transformers==3.3.1
uv>=0.5.0
uvicorn>=0.32.0
beautifulsoup4>=4.12.3
prometheus-client>=0.21.0
google-generativeai>=0.8.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.150.0
google-cloud-aiplatform>=1.70.0
tqdm>=4.66.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
unstructured>=0.10.0
